+++ lib/petal_components/showcase/registry.ex
@@ -47,6 +47,7 @@ defmodule PetalComponents.Showcase.Registry do
     PetalComponents.Showcase.ShineBorder,
     PetalComponents.Showcase.Skeleton,
     PetalComponents.Showcase.SlideOver,
+    PetalComponents.Showcase.Sortable,
     PetalComponents.Showcase.Sparkline,
     PetalComponents.Showcase.SpotlightCard,
     PetalComponents.Showcase.Stepper,